Castling Fix
Changes made to ensure that promotion happened on the turn of the player that earned it (i.e. when the camera was still in their perspective) broke Castling. Those changes have been undone & a better way for fixing promotion to be done on the correct turn needs to be implemented.